+++ utils.go
@@ -4,8 +4,11 @@ import (
 	"bufio"
 	"bytes"
 	"fmt"
+	"io/ioutil"
 	"log"
 	"os/exec"
+	"regexp"
+	"strconv"
 	"strings"
 	"time"
 )
@@ -17,11 +20,9 @@ var (
 // sh is a simple os.exec Command tool, returns trimmed string output
 func sh(name string, args ...string) (string, error) {
 	cmd := exec.Command(name, args...)
-	if isDebugEnabled() {
-		log.Printf("DEBUG: sh CMD: %q", cmd)
-	}
-	// TODO: capture and output STDERR to logfile?
+	log.Printf("INFO: sh CMD: %q", cmd)
 	out, err := cmd.Output()
+	log.Printf("INFO: [out, err]/[%s, %s]", out, err)
 	return strings.Trim(string(out), " \n"), err
 }
 
@@ -92,4 +93,76 @@ func grepLines(data string, like string) []string {
 	}
 
 	return result
+}
+
+// regexpLines pulls out lines that match a regexp as group matches
+func regexpLines(data string, regexp_s string) [][]string {
+	var result = [][]string{}
+
+	r, err := regexp.Compile(regexp_s)
+	if err != nil {
+		log.Printf("ERROR: unable to compile regexp")
+		return result
+	}
+
+	scanner := bufio.NewScanner(strings.NewReader(data))
+	for scanner.Scan() {
+		s := scanner.Text()
+		if r.MatchString(s) {
+			result = append(result, r.FindStringSubmatch(s))
+		}
+	}
+	if err := scanner.Err(); err != nil {
+		log.Printf("WARN: error scanning string for %s: %s", regexp_s, err)
+	}
+
+	return result
+}
+
+// Linux process management
+//ref: https://github.com/kimpettersen/GoProcs/blob/master/src/procs.go
+type Process struct {
+	Pid        string
+	Executable string
+}
+
+// List all processes in system
+func listProcesses() []Process {
+	var processes []Process
+	files, err := ioutil.ReadDir("/proc")
+
+	if err != nil {
+		log.Fatal("Could not read dir /proc :)")
+	}
+	var proc Process
+
+	for _, file := range files {
+		if _, err := strconv.Atoi(file.Name()); err == nil {
+
+			cmd, err := ioutil.ReadFile("/proc/" + file.Name() + "/cmdline")
+
+			cmdString := strings.Join(strings.Split(string(cmd), "\x00"), " ")
+
+			if err != nil {
+				log.Fatal("Can't read file:", err)
+			}
+
+			proc = Process{
+				Pid:        file.Name(),
+				Executable: cmdString,
+			}
+			processes = append(processes, proc)
+		}
+	}
+	return processes
+}
+
+// kill a process
+func kill(proc Process, signal string) error {
+	if err := exec.Command("kill", "-"+signal, string(proc.Pid)).Start(); err != nil {
+		log.Fatal("Couldn't kill ", err)
+		return err
+	}
+	fmt.Printf("Killed %v\n", proc.Executable)
+	return nil
 }